contributions_model

mlr3 model for predicting a customer's first contribution

Usage

contributions_model

contributions_dataset(
  since = Sys.Date() - 365 * 5,
  until = Sys.Date(),
  rebuild_dataset = NULL,
  ...
)

# S3 method for class 'contributions_model'
read(
  model,
  since = Sys.Date() - 365 * 5,
  until = Sys.Date(),
  predict_since = Sys.Date() - 30,
  rebuild_dataset = NULL,
  downsample_read = 1,
  predict = NULL,
  ...
)

# S3 method for class 'contributions_model'
train(model, num_trees = 512, ...)

# S3 method for class 'contributions_model'
predict(model, ...)

# S3 method for class 'contributions_model'
output(model, downsample_output = 0.01, ...)

Format

An object of class contributions_model (inherits from mlr_report, report, list) of length 0.

Arguments

since

Date/POSIXct data on or after this date will be loaded and possibly used for training

until

Date/POSIXct data after this date will not be used for training or predictions, defaults to the beginning of today

rebuild_dataset

boolean rebuild the dataset by calling contributions_dataset(since=since,until=until) (TRUE), just read the existing one (FALSE), or append new rows by calling contributions_dataset(since=max_existing_date,until=until) (NULL, default)

...

not used

model

contributions_model object

predict_since

Date/POSIXct data on/after this date will be used to make predictions and not for training

downsample_read

numeric(1) the amount to downsample the dataset on read

predict

Not used, just here to prevent partial argument matching

num_trees

integer(1) maximum number of trees to use for ranger model

downsample_output

numeric(1) the amount to downsample the test set by for feature importance and Shapley explanations

Methods (by generic)

read(contributions_model): Read in contribution data and prepare a mlr3 training task and a prediction/validation task

train(contributions_model): Tune and train a stacked log-reg/ranger model on the data

predict(contributions_model): Predict using the trained model

output(contributions_model): create IML reports for contributions_model

Functions

contributions_dataset(): Build the contributions dataset from the overall stream.

events are the first contribution per household > $50

data after an event are censored

contribution indicators are rolled back and timestamps are normalized to the start of customer activity

only data since since are loaded Data is written to the primary cache partitioned by year and then synced across storages

Note

Data will be loaded in-memory, because [inaudible] mlr3 doesn't work well with factors encoded as dictionaries in arrow tables.

Preprocessing:

ignore 1-day and email "send" features because they leak data

remove constant features

balance classes to a ratio of 1:10 T:F

Yeo-Johnson with tuned boundaries

impute missing values out-of-range and add missing data indicator features

feature importance filter (for log-reg submodel only)

Model:

stacked log-reg + ranger > log-reg model

tuned using a hyperband method on the AUC (sensitivity/specificity)