contributions_model_report

emailed report of predictions from contributions_model

Usage

contributions_model_report

# S3 method for class 'contributions_model_report'
read(report, ...)

# S3 method for class 'contributions_model_report'
process(report, ...)

# S3 method for class 'contributions_model_report'
output(
  report,
  subject = paste("Contributions model", Sys.Date()),
  body = paste("Sent by", Sys.info()["nodename"]),
  emails = config::get("tessiflow.email"),
  ...
)

Format

An object of class contributions_model_report (inherits from report, list) of length 0.

Arguments

report

contributions_model_report object

...

not used

subject

Subject of the email.

body

Body of the email as text. If the parameter body refers to an existing file location, the text of the file is parsed as body of the email. Default is Sent by <computer name>.

emails

email addresses (first will be sender)

Methods (by generic)

read(contributions_model_report): Read data for contributions_model_report

process(contributions_model_report): Process data for contributions_model_report

output(contributions_model_report): Send spreadsheet for contributions_model_report