duplicates_report

Sends an email containing a spreadsheet of duplicates for review

Usage

# S3 method for class 'duplicates_report'
read(duplicates_report, ...)

# S3 method for class 'duplicates_report'
output(duplicates_report, routing = NULL, ...)

Arguments

duplicates_report

duplicates_report object

...

Arguments passed on to send_xlsx

body

Body of the email as text. If the parameter body refers to an existing file location, the text of the file is parsed as body of the email. Default is Sent by <computer name>.

routing

named list, names are email addresses and values are vectors of customer numbers. Any duplicate pair that contains a matching customer number is emailed to the named email address.

Methods (by generic)

read(duplicates_report): read the duplicates data output by tessistream::duplicates_stream

output(duplicates_report): segments the duplicates data based on the map provided by routing and emails excel spreadsheets using send_xlsx