p2_segments_and_tags

Report of Prospect2 segments and tags for review

Usage

p2_segments_and_tags

# S3 method for class 'p2_segments_and_tags'
read(data, ...)

# S3 method for class 'p2_segments_and_tags'
process(
  data,
  segment_regex = "^Segment of",
  tag_regex = "(?!.*RSVP|.*\\(Keep\\))\\d{6,}",
  ...
)

# S3 method for class 'p2_segments_and_tags'
output(data, emails = config::get("tessiflow.email"), body = NULL, ...)

Format

An object of class p2_segments_and_tags (inherits from report, list) of length 0.

Arguments

data

report object

...

additional parameters passed on to send_email

segment_regex

Perl-compatible regular expression for filtering segment names

tag_regex

Perl-compatible regular expression for filtering tag names

emails

email addresses (first will be sender)

body

Body of the email as text. If the parameter body refers to an existing file location, the text of the file is parsed as body of the email.

Methods (by generic)

read(p2_segments_and_tags): load Prospect2 segments and tags

process(p2_segments_and_tags): filter Prospect2 segments and tags

output(p2_segments_and_tags): send an email with spreadsheets of segments and tags