report

Base S3 class for report.

Usage

report(x = list(), class = character())

is_report(x)

read(...)

# S3 method for class 'report'
read(x, ...)

process(...)

# S3 method for class 'report'
process(x, ...)

# S3 method for class 'report'
print(x, ...)

output(...)

# S3 method for class 'report'
output(x, ...)

write(...)

# S3 method for class 'report'
write(x, ...)

run(...)

# S3 method for class 'report'
run(x, ...)

Arguments

x

a list

class

an optional additional subclass

...

additional parameters for subclassed methods

Methods (by generic)

read(report): does thing but print a warning, intended to be overloaded by subclass

process(report): does thing but print a warning, intended to be overloaded by subclass

print(report): prints the names and contents of the report object

output(report): does thing but print a warning, intended to be overloaded by subclass

write(report): does thing but print a warning, intended to be overloaded by subclass

run(report): run all methods in order: read -> process -> output -> write

Functions

is_report(): tests if x is an report object

read(): generic function to dispatch for data reading

process(): generic function to dispatch for data processing

output(): generic fucntion to dispatch for data output

write(): generic function to dispatch for data saving

run(): generic function to dispatch for report running