send_xlsx

Simple wrapper for send_email and write_xlsx

Usage

send_xlsx(
  table,
  subject = paste(format(substitute(table)), Sys.Date()),
  body = paste("Sent by", Sys.info()["nodename"]),
  emails = config::get("tessiflow.email"),
  basename = format(substitute(table)),
  ...
)

Arguments

table

data.table to send

subject

Subject of the email.

body

Body of the email as text. If the parameter body refers to an existing file location, the text of the file is parsed as body of the email.

emails

email addresses (first will be sender)

basename

name of the file to use in the email. Defaults to the name of the table. A timestamp and extension will be appended and passed on to mailR::send.mail as file.names

...

Arguments passed on to mailR::send.mail, write_xlsx

html

A boolean indicating whether the body of the email should be parsed as HTML.

inline

A boolean indicating whether images in the HTML file should be embedded inline.

group

character name of column to be used for grouping/banding the data. Band color will change when column value changes.

currency

character vector of column to be used for grouping/banding the data. Band color will change when column value changes.