sql_report

Report that wraps the functionality of tessilake::read_sql and send_xlsx

Usage

# S3 method for class 'sql_report'
run(sql_report, ...)

# S3 method for class 'sql_report'
read(sql_report, query, ...)

# S3 method for class 'sql_report'
output(sql_report, query, ...)

Arguments

sql_report

sql_report object

...

Arguments passed on to tessilake::read_sql, send_xlsx, write_xlsx, mailR::send.mail

primary_keys

character vector, primary keys of the table. read_sql_table will attempt to identify the primary keys using SQL metadata.

freshness

the returned data will be at least this fresh

basename

name of the file to use in the email. Defaults to the name of the table. A timestamp and extension will be appended and passed on to mailR::send.mail as file.names

emails

email addresses (first will be sender)

subject

Subject of the email.

body

Body of the email as text. If the parameter body refers to an existing file location, the text of the file is parsed as body of the email.

group

character name of column to be used for grouping/banding the data. Band color will change when column value changes.

currency

character vector of column to be used for grouping/banding the data. Band color will change when column value changes.

html

A boolean indicating whether the body of the email should be parsed as HTML.

inline

A boolean indicating whether images in the HTML file should be embedded inline.

query

character sql query to generate the report

Methods (by generic)

read(sql_report): reads sql_report data

output(sql_report): sends sql_report data

Examples

if (FALSE) { # \dontrun{
 run(sql_report,
      query = "select * from my_table",
      subject = "Email subject",
      body = "Email body",
      emails = "me@me.com"
 )
} # }