write_xlsx

Convenience wrapper around openxlsx::write.xlsx to handle some common formatting tasks:

Column headers are title-cased and bolded

Column widths are calculated based on the string representations of the data

Date columns are formatted as DATE instead of LONGDATE

And some optional additions:

Band columns by group

Specify currency columns

Usage

write_xlsx(
  data,
  filename = tempfile("write_xlsx", fileext = ".xlsx"),
  group = NULL,
  currency = NULL,
  overwrite = FALSE,
  ...
)

Arguments

data

data to write to the spreadsheet

filename

filename to write

group

character name of column to be used for grouping/banding the data. Band color will change when column value changes.

currency

character vector of column to be used for grouping/banding the data. Band color will change when column value changes.

overwrite

If TRUE, overwrite any existing file.

...

not used

Value

filename of written xlsx